Indian Liver Patient Dataset

This data set contains 416 liver patient records and 167 non liver patient records. The data set was collected from north east of Andhra Pradesh, India. Selector is a class label used to divide into groups(liver patient or not). This data set contains 441 male patient records and 142 female patient records.

ILPD

Format

A data frame with 583 rows and 11 variables:

Age

Gender

Total_Bilirubin

Direct_Bilirubin

Alkaline_Phosphatase

Alamine_Aminotransferase

Aspartate_Aminotransferase

Total_Protein

Albumin

Albumin_Globulin_Ratio

Case_or_Control

Selector field used to split the data into two sets (labeled by the experts)

Source

https://bit.ly/3NjtlSe

Details

1. Bendi Venkata Ramana, Prof. M. S. Prasad Babu and Prof. N. B. Venkateswarlu, A Critical Comparative Study of Liver Patients from USA and INDIA: An Exploratory Analysis, International Journal of Computer Science Issues, ISSN :1694-0784, May 2012. 2. Bendi Venkata Ramana, Prof. M. S. Prasad Babu and Prof. N. B. Venkateswarlu, A Critical Study of Selected Classification Algorithms for Liver Disease Diagnosis, International Journal of Database Management Systems (IJDMS), Vol.3, No.2, ISSN : 0975-5705, PP 101-114, May 2011. 3. Dua, D. and Graff, C. (2019). UCI Machine Learning Repository [http://archive.ics.uci.edu/ml]. Irvine, CA: University of California, School of Information and Computer Science.