dtrackr: Track your Data Pipelines

Track and document 'dplyr' data pipelines. As you filter, mutate, and join your way through a data set, 'dtrackr' seamlessly keeps track of your data flow and makes publication ready documentation of a data pipeline simple.

See also

Author

Maintainer: Robert Challen rob.challen@bristol.ac.uk (ORCID)