Filtering data

Filter acts in the same way as in dplyr where predicates which evaluate to TRUE act to select items to include, and items for which the predicate cannot be evaluated are excluded. For tracking prior to the filter operation the size of each group is calculated {.count.in} and after the operation the output size of each group {.count.out}. The grouping {.strata} is also available (if grouped) for reporting. See dplyr::filter().

# S3 method for trackr_df
filter(
  .data,
  ...,
  .preserve = FALSE,
  .messages = "excluded {.excluded} items",
  .headline = .defaultHeadline(),
  .type = "exclusion",
  .asOffshoot = (.type == "exclusion"),
  .stage = "",
  .tag = NULL
)

Arguments

.data
a dataframe which may be grouped

...
the filter criteria

.preserve
Relevant when the .data input is grouped. If .preserve = FALSE (the default), the grouping structure is recalculated based on the resulting data, otherwise the grouping is kept as is.

.messages
a set of glue specs. The glue code can use any global variable, or {.strata},{.count.in},and {.count.out}

.headline
a headline glue spec. The glue code can use any global variable, or {.strata},{.count.in},and {.count.out}

.type
the format type of the action typically an exclusion

.asOffshoot
if the type is exclusion, .asOffshoot places the information box outside of the main flow, as an exclusion.

.stage
a name for this step in the pathway

.tag
if you want the summary data from this step in the future then give it a name with .tag.

Value

the filtered .data dataframe with history graph updated

See also

Examples

library(dplyr)
library(dtrackr)
tmp = iris %>% track() %>% group_by(Species)
tmp %>% filter(Petal.Length > 5) %>% history()

#> dtrackr history:
#> number of flowchart steps: 2 (approx)
#> tags defined: <none>
#> items excluded so far: <not capturing exclusions>
#> last entry / entries:
#> └ "stratify by Species"