Include any items matching a criteria

Apply a set of inclusion criteria and dplyr::summarise the actions of the filter to the dtrackr history graph Because of the ... filter specification, all parameters MUST BE NAMED. The criteria work in an alternative manner, i.e. the results include anything that match any of the criteria. If na.rm = TRUE they also keep anything that cannot be evaluated by a criteria - that may be true.

include_any(
  .data,
  ...,
  .headline = .defaultHeadline(),
  na.rm = TRUE,
  .type = "inclusion",
  .asOffshoot = FALSE
)

Arguments

.data

- a dataframe which may be grouped

...

- a dplyr filter specification as a formula where the RHS is a glue specification, defining the message. This can refer to grouping variables, variables from the environment and {.included} and {.matched} or {.missing} (included = matched+missing), {.count} and {.total} - group and overall counts respectively, e.g. "excluding {.matched} items and {.missing} with missing values".

.headline

- a glue specification which can refer to grouping variables of .data, or any variables defined in the calling environment

na.rm

- (default FALSE) if the filter cannot be evaluated for a row count that row as missing and either exclude it (TRUE) or don't exclude it (FALSE)

.type

- default "exclusion": used to define formatting

.asOffshoot

- do you want this comment to be an offshoot of the main flow (default = TRUE).

Value

the filtered .data dataframe with the history graph updated

Examples

iris %>%
   p_include_any(
      Petal.Length > 5 ~ "{.included} long ones",
      Petal.Length < 2 ~ "{.included} short ones"
   ) %>%
   p_get()
#> $excluded
#> NULL
#> 
#> $capture
#> [1] FALSE
#> 
#> $tags
#> # A tibble: 0 x 2
#> # … with 2 variables: .tag <chr>, .content <list>
#> 
#> $nodes
#> # A tibble: 1 x 6
#>     .id .rank .strata .label                                      .type   .stage
#>   <dbl> <dbl> <chr>   <chr>                                       <chr>   <chr> 
#> 1     1     1 ""      inclusions:<BR ALIGN='LEFT'/>42 long ones<… inclus… ""    
#> 
#> $edges
#> # A tibble: 0 x 5
#> # … with 5 variables: .to <dbl>, .from <int>, .rel <chr>, .strata <chr>,
#> #   .strata.prev <chr>
#> 
#> $head
#> # A tibble: 1 x 2
#>   .from .strata
#>   <dbl> <chr>  
#> 1     1 ""     
#>