Filtering data

Filter acts in the same way as DPLYR. Prior to the operation the size of the group is calculated {.count.in} and after the operation the output size {.count.out} The group {.strata} is also available (if grouped) for reporting

p_filter(
  .data,
  ...,
  .preserve = FALSE,
  .messages = "excluded {.excluded} items",
  .headline = .defaultHeadline(),
  .type = "exclusion",
  .asOffshoot = (.type == "exclusion"),
  .stage = "",
  .tag = NULL
)

Arguments

.data

- a dataframe which may be grouped

...

the filter criteria

.preserve

Relevant when the .data input is grouped. If .preserve = FALSE (the default), the grouping structure is recalculated based on the resulting data, otherwise the grouping is kept as is.

.messages

- a set of glue specs. The glue code can use any global variable, or {.strata},{.count.in},and {.count.out}

.headline

- a headline glue spec. The glue code can use any global variable, or {.strata},{.count.in},and {.count.out}

.type

- the format type of the action - typically an exclusion

.asOffshoot

- if the type is exclusion, asOffshoot places the information box outside of the main flow, as an exclusion.

.stage

- a name for this step in the pathway

.tag

- if you want the summary data from this step in the future then give it a name with .tag.

Value

the transformed .data.

Examples

iris %>% p_group_by(Species) %>% p_filter(Petal.Length > 5) %>% p_get()
#> $excluded
#> NULL
#> 
#> $capture
#> [1] FALSE
#> 
#> $tags
#> # A tibble: 0 x 2
#> # … with 2 variables: .tag <chr>, .content <list>
#> 
#> $nodes
#> # A tibble: 4 x 6
#>     .id .rank .strata       .label                                 .type  .stage
#>   <dbl> <dbl> <chr>         <chr>                                  <chr>  <chr> 
#> 1     1     1 ""            stratify by Species<BR ALIGN='LEFT'/>  strat… ""    
#> 2     2     1 "Species:set… <B>Species:setosa</B><BR ALIGN='LEFT'… exclu… ""    
#> 3     3     1 "Species:ver… <B>Species:versicolor</B><BR ALIGN='L… exclu… ""    
#> 4     4     1 "Species:vir… <B>Species:virginica</B><BR ALIGN='LE… exclu… ""    
#> 
#> $edges
#> # A tibble: 3 x 5
#>     .to .from .rel      .strata            .strata.prev
#>   <dbl> <dbl> <chr>     <chr>              <chr>       
#> 1     2     1 exclusion Species:setosa     ""          
#> 2     3     1 exclusion Species:versicolor ""          
#> 3     4     1 exclusion Species:virginica  ""          
#> 
#> $head
#> # A tibble: 1 x 2
#>   .from .strata
#>   <dbl> <chr>  
#> 1     1 ""     
#>