dplyr modifying operations

Equivalent dplyr functions for mutating, selecting and renaming a data set act in the normal way. mutates / selects / rename generally don't add anything in documentation so the default behaviour is to miss these out of the history. This can be overridden with the .messages, or .headline values in which case they behave just like a comment() See dplyr::mutate(), dplyr::add_count(), dplyr::add_tally(), dplyr::transmute(), dplyr::select(), dplyr::relocate(), dplyr::rename() dplyr::rename_with(), dplyr::arrange() for more details.

# S3 method for trackr_df
select(.data, ..., .messages = "", .headline = "", .tag = NULL)

Arguments

.data
A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr). See Methods, below, for more details.

...
<data-masking> Name-value pairs. The name gives the name of the column in the output.

The value can be:

A vector of length 1, which will be recycled to the correct length.

A vector the same length as the current group (or the whole data frame if ungrouped).

NULL, to remove the column.

A data frame or tibble, to create multiple columns in the output.

.messages
a set of glue specs. The glue code can use any global variable, grouping variable, {.new_cols} or {.dropped_cols} for changes to columns, {.cols} for the output column names, or {.strata}. Defaults to nothing.

.headline
a headline glue spec. The glue code can use any global variable, grouping variable, {.new_cols}, {.dropped_cols}, {.cols} or {.strata}. Defaults to nothing.

.tag
if you want the summary data from this step in the future then give it a name with .tag.

Value

the .data dataframe after being modified by the dplyr equivalent function, but with the history graph updated with a new stage if the .messages or .headline parameter is not empty.

See also

Examples

library(dplyr)
library(tidyselect)
library(tidyr)
library(dtrackr)

# mutate and other functions are unitary operations that generally change
# the structure but not size of a dataframe. In dtrackr these are by ignored
# by default but we can change that so that their behaviour is obvious.

# mutate
# In this example we compare the column names of the input and the
# output to identify the new columns created by the mutate operation as
# the `.new_cols` variable
iris %>%
  track() %>%
  mutate(extra_col = NA_real_,
         .messages="{.new_cols}",
         .headline="Extra columns from mutate:") %>%
  history()

#> dtrackr history:
#> number of flowchart steps: 2 (approx)
#> tags defined: <none>
#> items excluded so far: <not capturing exclusions>
#> last entry / entries:
#> └ "Extra columns from mutate:", "extra_col"

# transmute
# Here we do the same for a transmute()
iris %>%
  track() %>%
  group_by(Species) %>%
  transmute(
    sepal.w = Sepal.Width-1,
    sepal.l = Sepal.Length+1,
    .messages="{.new_cols}",
    .headline="New columns from transmute:") %>%
  history()

#> dtrackr history:
#> number of flowchart steps: 3 (approx)
#> tags defined: <none>
#> items excluded so far: <not capturing exclusions>
#> last entry / entries:
#> ├ [Species:setosa]: "New columns from transmute:", "sepal.w, sepal.l"
#> ├ [Species:versicolor]: "New columns from transmute:", "sepal.w, sepal.l"
#> └ [Species:virginica]: "New columns from transmute:", "sepal.w, sepal.l"

# select
# The output of the select verb (here using tidyselect syntax) can be captured
# and here all column names are being reported with the .cols variable.
iris %>%
  track() %>%
  group_by(Species) %>%
  select(
    starts_with("Sepal"),
    .messages="{.cols}",
    .headline="Output columns from select:") %>%
  history()

#> Adding missing grouping variables: `Species`

#> dtrackr history:
#> number of flowchart steps: 3 (approx)
#> tags defined: <none>
#> items excluded so far: <not capturing exclusions>
#> last entry / entries:
#> ├ [Species:setosa]: "Output columns from select:", "Species, Sepal.Length, Sepal.Width"
#> ├ [Species:versicolor]: "Output columns from select:", "Species, Sepal.Length, Sepal.Width"
#> └ [Species:virginica]: "Output columns from select:", "Species, Sepal.Length, Sepal.Width"

# relocate, this shows how the columns can be reordered
iris %>%
  track() %>%
  group_by(Species) %>%
  relocate(
    starts_with("Sepal"),
    .after=Species,
    .messages="{.cols}",
    .headline="Order of columns from relocate:") %>%
  history()

#> dtrackr history:
#> number of flowchart steps: 3 (approx)
#> tags defined: <none>
#> items excluded so far: <not capturing exclusions>
#> last entry / entries:
#> ├ [Species:setosa]: "Order of columns from relocate:", "Petal.Length, Petal.Width, Species, Sepal.Length, Sepal.Width"
#> ├ [Species:versicolor]: "Order of columns from relocate:", "Petal.Length, Petal.Width, Species, Sepal.Length, Sepal.Width"
#> └ [Species:virginica]: "Order of columns from relocate:", "Petal.Length, Petal.Width, Species, Sepal.Length, Sepal.Width"

# rename can show us which columns are new and which have been
# removed (with .dropped_cols)
iris %>%
  track() %>%
  group_by(Species) %>%
  rename(
    Stamen.Width = Sepal.Width,
    Stamen.Length = Sepal.Length,
    .messages=c("added {.new_cols}","dropped {.dropped_cols}"),
    .headline="Renamed columns:") %>%
  history()

#> dtrackr history:
#> number of flowchart steps: 3 (approx)
#> tags defined: <none>
#> items excluded so far: <not capturing exclusions>
#> last entry / entries:
#> ├ [Species:setosa]: "Renamed columns:", "added Stamen.Length, Stamen.Width", "dropped Sepal.Length, Sepal.Width"
#> ├ [Species:versicolor]: "Renamed columns:", "added Stamen.Length, Stamen.Width", "dropped Sepal.Length, Sepal.Width"
#> └ [Species:virginica]: "Renamed columns:", "added Stamen.Length, Stamen.Width", "dropped Sepal.Length, Sepal.Width"

# rename_with uses a function
iris %>%
  track() %>%
  group_by(Species) %>%
  rename_with(
    .fn = toupper,
    .messages="{.new_cols}",
    .headline="Renamed columns:") %>%
  history()

#> dtrackr history:
#> number of flowchart steps: 3 (approx)
#> tags defined: <none>
#> items excluded so far: <not capturing exclusions>
#> last entry / entries:
#> ├ [SPECIES:setosa]: "Renamed columns:", "SEPAL.LENGTH, SEPAL.WIDTH, PETAL.LENGTH, PETAL.WIDTH, SPECIES"
#> ├ [SPECIES:versicolor]: "Renamed columns:", "SEPAL.LENGTH, SEPAL.WIDTH, PETAL.LENGTH, PETAL.WIDTH, SPECIES"
#> └ [SPECIES:virginica]: "Renamed columns:", "SEPAL.LENGTH, SEPAL.WIDTH, PETAL.LENGTH, PETAL.WIDTH, SPECIES"

# arrange
# In this case we sort the data descending and show the first value
# is the same as the maximum value.
iris %>%
  track() %>%
  arrange(
    desc(Petal.Width),
    .messages="{.count} items, columns: {.cols}",
    .headline="Reordered dataframe:") %>%
  history()

#> dtrackr history:
#> number of flowchart steps: 2 (approx)
#> tags defined: <none>
#> items excluded so far: <not capturing exclusions>
#> last entry / entries:
#> └ "Reordered dataframe:", "150 items, columns: Sepal.Length, Sepal.Width, Petal.Length, Petal.Width, Species"

# add_count
# adding in a count or tally column as a new column
iris %>%
  track() %>%
  add_count(Species, name="new_count_total",
            .messages="{.new_cols}",
            # .messages="{.cols}",
            .headline="New columns from add_count:") %>%
  history()

#> dtrackr history:
#> number of flowchart steps: 2 (approx)
#> tags defined: <none>
#> items excluded so far: <not capturing exclusions>
#> last entry / entries:
#> └ "New columns from add_count:", "new_count_total"

# add_tally
iris %>%
  track() %>%
  group_by(Species) %>%
  dtrackr::add_tally(wt=Petal.Length, name="new_tally_total",
            .messages="{.new_cols}",
            .headline="New columns from add_tally:") %>%
  history()

#> dtrackr history:
#> number of flowchart steps: 3 (approx)
#> tags defined: <none>
#> items excluded so far: <not capturing exclusions>
#> last entry / entries:
#> ├ [Species:setosa]: "New columns from add_tally:", "new_tally_total"
#> ├ [Species:versicolor]: "New columns from add_tally:", "new_tally_total"
#> └ [Species:virginica]: "New columns from add_tally:", "new_tally_total"