Add a summary to the dtrackr history graph

Because of the ... summary specification parameters MUST BE NAMED.

status(
  .data,
  ...,
  .messages = NULL,
  .headline = NULL,
  .type = "info",
  .asOffshoot = FALSE
)

Arguments

.data
- a dataframe which may be grouped

...
- any normal dplyr::summarise specification, e.g. count=n() or av = mean(x) etc.

.messages
- a character vector of glue specifications. A glue specification can refer to the summary outputs, any grouping variables of .data, the .strata, or any variables defined in the calling environment

.headline
- a glue specification which can refer to grouping variables of .data, or any variables defined in the calling environment

.type
- one of "info","exclusion": used to define formatting

.asOffshoot
- do you want this comment to be an offshoot of the main flow (default = FALSE).

Value

the same .data dataframe with the history graph updated

Examples

iris %>% p_group_by(Species) %>%
   p_status(
      long = p_count_if(Petal.Length>5),
      short = p_count_if(Petal.Length<2),
      .messages="{Species}: {long} long ones & {short} short ones"
   ) %>% p_get()

#> $nodes
#> # A tibble: 4 x 5
#>     .id .rank .strata        .label                                       .type 
#>   <dbl> <dbl> <chr>          <chr>                                        <chr> 
#> 1     1     1 ""             stratify by Species<BR ALIGN='LEFT'/>        strat…
#> 2     2     2 "Species:seto… <B>Species:setosa</B><BR ALIGN='LEFT'/>seto… info  
#> 3     3     2 "Species:vers… <B>Species:versicolor</B><BR ALIGN='LEFT'/>… info  
#> 4     4     2 "Species:virg… <B>Species:virginica</B><BR ALIGN='LEFT'/>v… info  
#> 
#> $edges
#> # A tibble: 3 x 5
#>     .to .from .rel  .strata            .strata.prev
#>   <dbl> <dbl> <chr> <chr>              <chr>       
#> 1     2     1 info  Species:setosa     ""          
#> 2     3     1 info  Species:versicolor ""          
#> 3     4     1 info  Species:virginica  ""          
#> 
#> $head
#> # A tibble: 3 x 2
#>   .from .strata           
#>   <dbl> <chr>             
#> 1     2 Species:setosa    
#> 2     3 Species:versicolor
#> 3     4 Species:virginica 
#>