Summarise a data set

Summarising a data set acts in the normal way. Any columns resulting form the summary can be added to the history graph In the history this joins any stratified branches and acts as a specific type of p_summary, allowing you to generate some summary statistics about the un-grouped data. See dplyr::summarise().

# S3 method for trackr_df
summarise(
  .data,
  ...,
  .groups = NULL,
  .messages = "",
  .headline = "",
  .tag = NULL
)

Arguments

.data

a dataframe which may be grouped

...
a set of dplyr summary expressions.

.groups

Experimental lifecycle Grouping structure of the result.

.messages

a set of glue specs. The glue code can use any summary variable defined in the ... parameter, or any global variable, or {.strata}

.headline

a headline glue spec. The glue code can use any summary variable defined in the ... parameter, or any global variable, or {.strata}

.tag

if you want the summary data from this step in the future then give it a name with .tag.

Value

the .data dataframe summarised with the history graph updated showing the summarise operation as a new stage

See also

Examples

library(dplyr)
tmp = iris %>% group_by(Species)
tmp %>% summarise(avg = mean(Petal.Length), .messages="{avg} length") %>% history()

#> dtrackr history:
#> number of flowchart steps: 0 (approx)
#> tags defined: <none>
#> items excluded so far: <not capturing exclusions>