Remove a stratification from a data set

Un-grouping a data set logically combines the different arms. In the history this joins any stratified branches and acts as a specific type of p_summary, allowing you to generate some summary statistics about the un-grouped data

# S3 method for trackr_df
ungroup(
  x,
  ...,
  .messages = .defaultMessage(),
  .headline = .defaultHeadline(),
  .tag = NULL
)

Arguments

x

- a dataframe which may be grouped (why not .data?)

...

a set of dplyr summary expressions. If left blank a default of ".count=n()" will be filled in

.messages

- a set of glue specs. The glue code can use any summary variable defined in the ... parameter, or any global variable, or {.count}. the default is "total .count items"

.headline

- a headline glue spec. The glue code can use any summary variable defined in the ... parameter, or {.count}.

.tag

- if you want the summary data from this step in the future then give it a name with .tag.

Value

the .data but dplyr::ungrouped.

Examples

iris %>%
   p_group_by(Species) %>%
   p_comment("A test") %>%
   p_ungroup(
      my_avg = mean(Petal.Length),
      my_count = dplyr::n(),
      .messages="{my_count} items with {my_avg} petal length"
   ) %>%
   p_get()
#> $excluded
#> NULL
#> 
#> $capture
#> [1] FALSE
#> 
#> $tags
#> # A tibble: 0 x 2
#> # … with 2 variables: .tag <chr>, .content <list>
#> 
#> $nodes
#> # A tibble: 5 x 6
#>     .id .rank .strata        .label                               .type   .stage
#>   <dbl> <dbl> <chr>          <chr>                                <chr>   <chr> 
#> 1     1     1 ""             stratify by Species<BR ALIGN='LEFT'… strati… ""    
#> 2     2     2 "Species:seto… <B>Species:setosa</B><BR ALIGN='LEF… info    ""    
#> 3     3     2 "Species:vers… <B>Species:versicolor</B><BR ALIGN=… info    ""    
#> 4     4     2 "Species:virg… <B>Species:virginica</B><BR ALIGN='… info    ""    
#> 5     5     3 ""             150 items with 3.758 petal length<B… summar… ""    
#> 
#> $edges
#> # A tibble: 6 x 5
#>     .to .from .rel      .strata              .strata.prev        
#>   <dbl> <dbl> <chr>     <chr>                <chr>               
#> 1     2     1 info      "Species:setosa"     ""                  
#> 2     3     1 info      "Species:versicolor" ""                  
#> 3     4     1 info      "Species:virginica"  ""                  
#> 4     5     2 summarise ""                   "Species:setosa"    
#> 5     5     3 summarise ""                   "Species:versicolor"
#> 6     5     4 summarise ""                   "Species:virginica" 
#> 
#> $head
#> # A tibble: 1 x 2
#>   .from .strata
#>   <dbl> <chr>  
#> 1     5 ""     
#>