Filtering data

Filter acts in the same way as DPLYR. Prior to the operation the size of the group is calculated .count.in and after the operation the output size .count.out The group .strata is also available (if grouped) for reporting

# S3 method for trackr_df
filter(
  .data,
  ...,
  .preserve = FALSE,
  .messages = "excluded {.count.in-.count.out} items",
  .headline = "{.strata}",
  .type = "exclusion",
  .asOffshoot = (.type == "exclusion")
)

Arguments

.data
- a dataframe which may be grouped

...
the filter criteria

.preserve
Relevant when the .data input is grouped. If .preserve = FALSE (the default), the grouping structure is recalculated based on the resulting data, otherwise the grouping is kept as is.

.messages
- a set of glue specs. The glue code can use any global variable, or .strata,.count.in,and .count.out

.headline
- a headline glue spec. The glue code can use any global variable, or .strata,.count.in,and .count.out

.type
- the format type of the action - typically an exclusion

.asOffshoot
- if the type is exclusion, asOffshoot places the information box outside of the main flow, as an exclusion.

Value

the transformed .data.

Examples

iris %>% p_group_by(Species) %>% p_filter(Petal.Length > 5) %>% p_get()

#> $nodes
#> # A tibble: 4 x 5
#>     .id .rank .strata        .label                                      .type  
#>   <dbl> <dbl> <chr>          <chr>                                       <chr>  
#> 1     1     1 ""             stratify by Species<BR ALIGN='LEFT'/>       strati…
#> 2     2     2 "Species:seto… <B>Species:setosa</B><BR ALIGN='LEFT'/>exc… exclus…
#> 3     3     2 "Species:vers… <B>Species:versicolor</B><BR ALIGN='LEFT'/… exclus…
#> 4     4     2 "Species:virg… <B>Species:virginica</B><BR ALIGN='LEFT'/>… exclus…
#> 
#> $edges
#> # A tibble: 3 x 5
#>     .to .from .rel      .strata            .strata.prev
#>   <dbl> <dbl> <chr>     <chr>              <chr>       
#> 1     2     1 exclusion Species:setosa     ""          
#> 2     3     1 exclusion Species:versicolor ""          
#> 3     4     1 exclusion Species:virginica  ""          
#> 
#> $head
#> # A tibble: 1 x 2
#>   .from .strata
#>   <dbl> <chr>  
#> 1     1 ""     
#>