Stratifying your analysis

Grouping a data set acts in the normal way.

# S3 method for trackr_df
group_by(
  .data,
  ...,
  .add = FALSE,
  .drop = dplyr::group_by_drop_default(.data),
  .messages = "stratify by {.cols}",
  .headline = NULL
)

Arguments

.data
- a dataframe which may be grouped

...
a set of dplyr column expressions.

.add
When FALSE, the default, group_by() will override existing groups. To add to the existing groups, use .add = TRUE.

This argument was previously called add, but that prevented creating a new grouping variable called add, and conflicts with our naming conventions.

.drop
Drop groups formed by factor levels that don't appear in the data? The default is TRUE except when .data has been previously grouped with .drop = FALSE. See group_by_drop_default() for details.

.messages
- a set of glue specs. The glue code can use any global variable, or .cols which is the columns that are being grouped by.

.headline
- a headline glue spec. The glue code can use any global variable, or .cols.

Value

the .data but grouped.

Examples

iris %>% p_group_by(Species, .messages="stratify by {.cols}") %>% p_comment("{.strata}") %>% p_get()

#> $nodes
#> # A tibble: 4 x 5
#>     .id .rank .strata        .label                                       .type 
#>   <dbl> <dbl> <chr>          <chr>                                        <chr> 
#> 1     1     1 ""             stratify by Species<BR ALIGN='LEFT'/>        strat…
#> 2     2     2 "Species:seto… <B>Species:setosa</B><BR ALIGN='LEFT'/>Spec… info  
#> 3     3     2 "Species:vers… <B>Species:versicolor</B><BR ALIGN='LEFT'/>… info  
#> 4     4     2 "Species:virg… <B>Species:virginica</B><BR ALIGN='LEFT'/>S… info  
#> 
#> $edges
#> # A tibble: 3 x 5
#>     .to .from .rel  .strata            .strata.prev
#>   <dbl> <dbl> <chr> <chr>              <chr>       
#> 1     2     1 info  Species:setosa     ""          
#> 2     3     1 info  Species:versicolor ""          
#> 3     4     1 info  Species:virginica  ""          
#> 
#> $head
#> # A tibble: 3 x 2
#>   .from .strata           
#>   <dbl> <chr>             
#> 1     2 Species:setosa    
#> 2     3 Species:versicolor
#> 3     4 Species:virginica 
#>