Union of two or more data sets

This merges the history of 2 dataframes and binds the rows. It calculates the total number of resulting rows as .count.out

p_bind_rows(
  ...,
  .id = NULL,
  .messages = "{.count.out} in union",
  .headline = "Union"
)

Arguments

...

the data frames to bind

.id

Data frame identifier.

When .id is supplied, a new column of identifiers is created to link each row to its original data frame. The labels are taken from the named arguments to bind_rows(). When a list of data frames is supplied, the labels are taken from the names of the list. If no names are found a numeric sequence is used instead.

.messages

- a set of glue specs. The glue code can use any global variable, or {.count.out}

.headline

- a glue spec. The glue code can use any global variable, or {.count.out}

Value

the union data frame

Examples

p_bind_rows( iris %>% p_comment("one"), iris %>% p_comment("two") ) %>% p_get()
#> $excluded
#> NULL
#> 
#> $capture
#> [1] FALSE
#> 
#> $tags
#> # A tibble: 0 x 2
#> # … with 2 variables: .tag <chr>, .content <list>
#> 
#> $nodes
#> # A tibble: 3 x 6
#>     .id .rank .strata .label                                       .type  .stage
#>   <dbl> <dbl> <chr>   <chr>                                        <chr>  <chr> 
#> 1     1     1 ""      one<BR ALIGN='LEFT'/>                        info   ""    
#> 2     2     1 ""      two<BR ALIGN='LEFT'/>                        info   ""    
#> 3     3     2 ""      <B>Union</B><BR ALIGN='LEFT'/>300 in union<… combi… ""    
#> 
#> $edges
#> # A tibble: 2 x 5
#>     .to .from .rel    .strata .strata.prev
#>   <dbl> <dbl> <chr>   <chr>   <chr>       
#> 1     3     1 combine ""      ""          
#> 2     3     2 combine ""      ""          
#> 
#> $head
#> # A tibble: 1 x 2
#>   .from .strata
#>   <dbl> <chr>  
#> 1     3 ""     
#>