Stratifying your analysis

Grouping a data set acts in the normal way. When tracking a dataframe sometimes a `group_by()` operation will create a lot of groups. This happens for example if you are doing a `group_by()`, `summarise()` step that is aggregating data on a fine scale, e.g. by day in a timeseries. This is generally a terrible idea when tracking a dataframe as the resulting flowchart will have many many branches. `dtrackr` will detect this issue and pause tracking the dataframe with a warning. It is up to the user to the `resume()` tracking when the large number of groups have been resolved e.g. using a `dplyr::ungroup()`. This limit is configurable with `options("dtrackr.max_supported_groupings"=XX)`. The default is 16.

p_group_by(
  .data,
  ...,
  .add = FALSE,
  .drop = dplyr::group_by_drop_default(.data),
  .messages = "stratify by {.cols}",
  .headline = NULL,
  .tag = NULL,
  .maxgroups = .defaultMaxSupportedGroupings()
)

Arguments

.data
- a dataframe which may be grouped

...
a set of dplyr column expressions.

.add
When FALSE, the default, group_by() will override existing groups. To add to the existing groups, use .add = TRUE.

This argument was previously called add, but that prevented creating a new grouping variable called add, and conflicts with our naming conventions.

.drop
Drop groups formed by factor levels that don't appear in the data? The default is TRUE except when .data has been previously grouped with .drop = FALSE. See group_by_drop_default() for details.

.messages
- a set of glue specs. The glue code can use any global variable, or {.cols} which is the columns that are being grouped by.

.headline
- a headline glue spec. The glue code can use any global variable, or {.cols}.

.tag
- if you want the summary data from this step in the future then give it a name with .tag.

.maxgroups
- the maximum number of subgroups allowed before the tracking is paused.

Value

the .data but grouped.

Examples

iris %>% p_group_by(Species, .messages="stratify by {.cols}") %>% p_comment("{.strata}") %>% p_get()

#> $excluded
#> NULL
#> 
#> $capture
#> [1] FALSE
#> 
#> $tags
#> # A tibble: 0 × 2
#> # … with 2 variables: .tag <chr>, .content <list>
#> 
#> $nodes
#> # A tibble: 4 × 6
#>     .id .rank .strata              .label                           .type .stage
#>   <dbl> <dbl> <chr>                <chr>                            <chr> <chr> 
#> 1     1     1 ""                   stratify by Species<BR ALIGN='L… stra… ""    
#> 2     2     2 "Species:setosa"     <B>Species:setosa</B><BR ALIGN=… info  ""    
#> 3     3     2 "Species:versicolor" <B>Species:versicolor</B><BR AL… info  ""    
#> 4     4     2 "Species:virginica"  <B>Species:virginica</B><BR ALI… info  ""    
#> 
#> $edges
#> # A tibble: 3 × 5
#>     .to .from .rel  .strata            .strata.prev
#>   <dbl> <dbl> <chr> <chr>              <chr>       
#> 1     2     1 info  Species:setosa     ""          
#> 2     3     1 info  Species:versicolor ""          
#> 3     4     1 info  Species:virginica  ""          
#> 
#> $head
#> # A tibble: 3 × 2
#>   .from .strata           
#>   <dbl> <chr>             
#> 1     2 Species:setosa    
#> 2     3 Species:versicolor
#> 3     4 Species:virginica 
#>